fix: close new dev warnings on this branch
Three files gained new sign-compare/missing-field-initializer
warnings from #666 (LinearLighting rework) and #634 (Wind system),
both merged to dev after this branch's original sweep. Also reverts
the UnifiedWater.cpp null-check from the prior commit: fileName is
a fixed char[MAX_PATH] embedded in RE::TESFile, never a pointer, so
!file->fileName was dead/tautological code that itself produced a
new -Wpointer-bool-conversion warning.

+++ src/Features/LinearLighting.cpp
@@ -161,7 +161,7 @@ void LinearLighting::OnWorldRenderEnd(RE::RENDER_TARGET a_renderTarget)
 		return;
 
 	const auto targetIndex = static_cast<size_t>(a_renderTarget);
-	if (!sceneGammaDecodeCS || !globals::game::renderer || targetIndex >= Util::GetRenderTargetCount())
+	if (!sceneGammaDecodeCS || !globals::game::renderer || targetIndex >= static_cast<size_t>(Util::GetRenderTargetCount()))
 		return;
 
 	auto& target = globals::game::renderer->GetRuntimeData().renderTargets[targetIndex];
